Rename internal config to module config
