+++ lib/configs/module.js
