+++ lib/index.js
@@ -5,7 +5,7 @@ const ForLoop = require('./rules/for-loop');
 const NoArrowception = require('./rules/no-arrowception');
 const NoVar = require('./rules/no-var');
 const Recommended = require('./configs/recommended');
-const Internal = require('./configs/internal');
+const Module = require('./configs/module');
 const ScopeStart = require('./rules/scope-start');
 
 
@@ -15,7 +15,7 @@ const internals = {};
 module.exports = {
     configs: {
         recommended: Recommended,
-        internal: Internal
+        module: Module
     },
     rules: {
         'capitalize-modules': CapitalizeModules,
